LPS-179094 Extract ScreenNavigationEntry methods from ObjectDefinitionsStatesScreenNavigationCategory to ObjectDefinitionsStatesScreenNavigationEntry, implement getCategoryKey method in ObjectDefinitionsStatesScreenNavigationEntry and make ObjectDefinitionsStatesScreenNavigationCategory extends BaseObjectDefinitionsScreenNavigationCategory
