+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/object/definitions/frontend/taglib/servlet/taglib/ObjectDefinitionsStatesScreenNavigationCategory.java
@@ -15,97 +15,24 @@
 package com.liferay.object.web.internal.object.definitions.frontend.taglib.servlet.taglib;
 
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.list.type.service.ListTypeDefinitionService;
-import com.liferay.object.model.ObjectDefinition;
-import com.liferay.object.service.ObjectFieldSettingLocalService;
 import com.liferay.object.web.internal.object.definitions.constants.ObjectDefinitionsScreenNavigationEntryConstants;
-import com.liferay.object.web.internal.object.definitions.display.context.ObjectDefinitionsStateManagerDisplayContext;
-import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
-
-import java.util.Locale;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Carlos Montenegro
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=70",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=70",
+	service = ScreenNavigationCategory.class
 )
 public class ObjectDefinitionsStatesScreenNavigationCategory
-	extends BaseObjectDefinitionsScreenNavigationEntry
-	implements ScreenNavigationCategory {
+	extends BaseObjectDefinitionsScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
 		return ObjectDefinitionsScreenNavigationEntryConstants.
 			CATEGORY_KEY_STATE_MANAGER;
 	}
-
-	@Override
-	public String getEntryKey() {
-		return ObjectDefinitionsScreenNavigationEntryConstants.
-			ENTRY_KEY_STATE_MANAGER;
-	}
-
-	@Override
-	public String getJspPath() {
-		return "/object_definitions/object_definition/states.jsp";
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return _language.get(locale, "state-manager");
-	}
-
-	@Override
-	public boolean isVisible(User user, ObjectDefinition objectDefinition) {
-		return !objectDefinition.isUnmodifiableSystemObject();
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			new ObjectDefinitionsStateManagerDisplayContext(
-				httpServletRequest, _listTypeDefinitionService,
-				_objectDefinitionModelResourcePermission,
-				_objectFieldSettingLocalService));
-
-		super.render(httpServletRequest, httpServletResponse);
-	}
-
-	@Reference
-	private Language _language;
-
-	@Reference
-	private ListTypeDefinitionService _listTypeDefinitionService;
-
-	@Reference(
-		target = "(model.class.name=com.liferay.object.model.ObjectDefinition)"
-	)
-	private ModelResourcePermission<ObjectDefinition>
-		_objectDefinitionModelResourcePermission;
-
-	@Reference
-	private ObjectFieldSettingLocalService _objectFieldSettingLocalService;
 
 }
